Optimization pass: reduce code duplication and split oversized functions (~1100 lines net)
- SaveSystem.cpp: Split RegisterBuiltins() (425 lines) into 3 domain functions
  (RegisterCoreSerializers, RegisterPhysicsSerializers, RegisterGameplaySerializers);
  refactor SerializeWorld() with TrySerialize<T> template helper (1350→1274 lines)
- InputManager.cpp: Consolidate duplicated Windows/Linux code — shared key mapping
  tables, query methods, and console integration defined once (1338→840 lines, -37%)
- UpscalingSystem.cpp: Extract ~640 lines of inline HLSL shader strings to new
  UpscalingShaders.h header (1335→652 lines, -51%)
- ConsoleApp.cpp: Split oversized Run/ReadEngineInput/ReadUserInput into focused
  private helpers (functions now under 60-line guideline)
- SparkEngine.cpp: Refactor RunHeadlessWindows to reuse existing helper functions
